+++ example/crash/game.dart
@@ -1,6 +1,7 @@
 library crash;
 
 import 'dart:html';
+import 'package:boarding/grid.dart';
 import 'package:boarding/pieces.dart';
 import 'package:boarding/boarding.dart';
 import 'package:boarding/util.dart';