+++ example/drop/game.dart
@@ -1,9 +1,9 @@
 library drop;
 
 import 'dart:html';
+import 'package:boarding/grid.dart';
 import 'package:boarding/pieces.dart';
 import 'package:boarding/boarding.dart';
-import 'package:boarding/util.dart';
 
 part 'view/board.dart';