+++ lib/pieces.dart
@@ -1,6 +1,7 @@
 library pieces;
 
 import 'dart:convert';
+import 'package:boarding/grid.dart';
 import 'package:boarding/util.dart';
 
 part 'model/pieces/pieces.dart';