+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 based on [Semantic Versioning](http://semver.org/)
 
+*0.2.1* 2015-03-09
+
++ move the accident method from RedCar to MovablePiece in the crash example
++ remove the FallingPiece class in pieces
++ remove the FallingPieces class in pieces
++ replace the move method with move([Direction direction]) of MovablePiece in pieces
++ replace FallingPiece with MovablePiece in Board of the drop example
+
 *0.2.1* 2015-03-08
 
 + add drawPiece method in Surface